+++ audio_service/lib/audio_service.dart
@@ -1317,6 +1317,7 @@ class AudioService {
     bool androidNotificationOngoing = false,
     bool androidResumeOnClick = true,
     bool androidStopForegroundOnPause = false,
+    bool androidStopForegroundOnCompleted = false,
     bool androidEnableQueue = false,
     Size? androidArtDownscaleSize,
     Duration fastForwardInterval = const Duration(seconds: 10),
@@ -1358,6 +1359,7 @@ class AudioService {
               androidNotificationClickStartsActivity,
           androidNotificationOngoing: androidNotificationOngoing,
           androidStopForegroundOnPause: androidStopForegroundOnPause,
+          androidStopForegroundOnCompleted: androidStopForegroundOnCompleted,
           artDownscaleWidth: androidArtDownscaleSize?.width.round(),
           artDownscaleHeight: androidArtDownscaleSize?.height.round(),
           fastForwardInterval: fastForwardInterval,
@@ -3439,6 +3441,13 @@ class AudioServiceConfig {
   /// able to kill your service at any time to reclaim resources.
   final bool androidStopForegroundOnPause;
 
+  /// Similar to [androidStopForegroundOnPause] except on occurs when the
+  /// AudioProcessingState == AudioProcessingState.completed. The Android service will
+  /// switch to a lower priority state allowing the user to swipe away the notification.
+  /// Note that while in this lower priority state, the operating system will also be
+  /// able to kill your service at any time to reclaim resources.
+  final bool androidStopForegroundOnCompleted;
+
   /// If not null, causes the artwork specified by [MediaItem.artUri] to be
   /// downscaled to this maximum pixel width. If the resolution of your artwork
   /// is particularly high, this can help to conserve memory. If specified,
@@ -3482,6 +3491,7 @@ class AudioServiceConfig {
     this.androidNotificationClickStartsActivity = true,
     this.androidNotificationOngoing = false,
     this.androidStopForegroundOnPause = true,
+    this.androidStopForegroundOnCompleted = true,
     this.artDownscaleWidth,
     this.artDownscaleHeight,
     this.fastForwardInterval = const Duration(seconds: 10),
@@ -3507,6 +3517,7 @@ class AudioServiceConfig {
             androidNotificationClickStartsActivity,
         androidNotificationOngoing: androidNotificationOngoing,
         androidStopForegroundOnPause: androidStopForegroundOnPause,
+        androidStopForegroundOnCompleted: androidStopForegroundOnCompleted,
         artDownscaleWidth: artDownscaleWidth,
         artDownscaleHeight: artDownscaleHeight,
         fastForwardInterval: fastForwardInterval,