+++ audio_service/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.18.13
 
 * Fix setAndroidPlaybackInfo call blocking (@julianscheel).
+* Pass through missing extras to playFrom.../prepareFrom... (@Ruchit2759).
 
 ## 0.18.12